+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Replaced `async-tungstenite` dependency with `tokio-tungstenite`
+- Added public `tungstenite` export
 - Bumped minimum supported Rust version to `1.43`
 - Bumped `tokio` dependency to `1.0`